Fix the MCM SDK build configuration to support the FFmpeg build system
To build FFmpeg with an MCM plugin, a .pc file has to be provided
along with the libmcm_dp.so library file. This is a requirement of the
FFmpeg build system that uses pkg-config to locate an external library.

This commit adds a template pkg-config file libmcm_dp.pc.in and
changes the cmake configuration to create a libmcm_dp.pc file from the
template and then installs it.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>