Send reply to Agent before applying the config in Media Proxy
When the ApplyConfig command is received in Proxy, the config is
parsed and immediately applied. After that, a reply is sent to Agent.
This flow brings a potential race condition in Proxy when SDK
requests to delete a connection and simultaneously Agent sends
the ApplyConfig command.

Change the flow to avoid mutual nested locking by replying to Agent
and then applying the received configuration.

Fix wrong indentation in the changed source file.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>