Reject requests to link multiple transmitters to group in Agent
By design, multipoint groups can have many receivers but only
up to one transmitter. Multiple Local Rx entities (Tx connections
or ingress bridges) should not be allowed in a multipoint group.

Before adding a connection to a multipoint group, check if
the group already has one Tx connected, and if so, reject the
request to link the connection to the group.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>