Fix issues in 25.03-rc3 (#347)
* Fix proxy star interconnection algorithm in Agent

Multiple RDMA bridges are created to pass traffic of the same
multipoint group between two nodes. Rework this to create only
one pair of RDMA bridges no matter how many receivers are
in the group.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Send reply to Agent before applying the config in Media Proxy

When the ApplyConfig command is received in Proxy, the config is
parsed and immediately applied. After that, a reply is sent to Agent.
This flow brings a potential race condition in Proxy when SDK
requests to delete a connection and simultaneously Agent sends
the ApplyConfig command.

Change the flow to avoid mutual nested locking by replying to Agent
and then applying the received configuration.

Fix wrong indentation in the changed source file.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Fix -a argument handling in Media Proxy

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Implement graceful shutdown handling in FFmpeg MCM plugin

Register a SIGINT/SIGTERM handler and in case of termination,
release the resources appropriately by deleting the connection
and then deleting the client.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Reject requests to link multiple transmitters to group in Agent

By design, multipoint groups can have many receivers but only
up to one transmitter. Multiple Local Rx entities (Tx connections
or ingress bridges) should not be allowed in a multipoint group.

Before adding a connection to a multipoint group, check if
the group already has one Tx connected, and if so, reject the
request to link the connection to the group.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Implement Proxy-Agent connection recovery mechanism

Requirements are as below:
* Media Proxy should continuously try to connect to Agent.
* If connection between Media Proxy and Agent breaks,
Media Proxy should reestablish the connection and avoid
affecting currently running streams.
* If Agent restarts, Media Proxy should reconnect to Agent.
Media Proxy should delete all existing connections.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

---------

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>