Add workaround to avoid first frames losses
When the user calls mesh_create_tx_connection(), SDK requests
Media Proxy to create a connection and register it in Mesh Agent.
The Agent then changes a configuration for all Media Proxies
and applies it. After that Media Proxies apply their configurations,
which results in creating of multipoint groups and bridges.
This whole process can take some time, so the first frames that
the user starts sending to the connection by calling mesh_put_buffer()
can be lost.

To avoid the frame losses, wait for 100 ms or so after creating
the Tx connection, and then return success to the user.

This should allow Mesh Agent and Media Proxies to complete
creation of multipoint groups and bridges after registering the
connection.

Make the delay configurable in the connection JSON config:
"connCreationDelayMilliseconds: 100"

The default delay is set to 0 ms to avoid adding latency by default.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>