Fix issues appeared in version 25.03-rc4 (#353)
* Add version info to Agent, Media Proxy, and SDK

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Increase string storage for JSON config in FFmpeg MCM plugin

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Improve shutdown flow in Media Proxy

Add a mechanism to avoid the race condition when unregistering
Media Proxy and exiting the command queue loop.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Add workaround to avoid first frames losses

When the user calls mesh_create_tx_connection(), SDK requests
Media Proxy to create a connection and register it in Mesh Agent.
The Agent then changes a configuration for all Media Proxies
and applies it. After that Media Proxies apply their configurations,
which results in creating of multipoint groups and bridges.
This whole process can take some time, so the first frames that
the user starts sending to the connection by calling mesh_put_buffer()
can be lost.

To avoid the frame losses, wait for 100 ms or so after creating
the Tx connection, and then return success to the user.

This should allow Mesh Agent and Media Proxies to complete
creation of multipoint groups and bridges after registering the
connection.

Make the delay configurable in the connection JSON config:
"connCreationDelayMilliseconds: 100"

The default delay is set to 0 ms to avoid adding latency by default.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Add support for conn creation delay in FFmpeg MCM plugin

Argument name is conn_delay, value is integer number of
milliseconds.

Note: the delay default value is 0 (no delay). The user can specify
some 50-100 ms delay to avoid first frames losses if needed.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Fix last Tx packet len in FFmpeg MCM plugin audio device

Use mesh_buffer_set_payload_len() to set the actual length of the
last packet of an audio transmission.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Add support for ST2110 payload type (96..127)

Modify SDK, Media Proxy, Mesh Agent, and FFmpeg plugin
to support SMPTE ST2110 payload type, which is typically 96..127.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Fix build system to support version info in separate SW components

There is an issue in CI/CD that sdk/build.sh couldn't find
mcm-version.h. Rework the build system to create mcm-version.*
in sdk, media-proxy, and control-plane-agent directories.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

---------

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>