Fix connection shutdown handling flow (#359)
* Implement lock-free hotpath in conn and multipoint group

Replace mutex-protected link storages in Connection and
MultipointGroup classes with a lock-free custom atomic
class DataplaneAtomicUint64.

This avoids mutex availability timeout issues happened in audio
transmission when shutting down a connection.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Fix FFmpeg MCM plugin Rx connection shutdown related issues

* Add FFmpeg MCM plugin argument to set the memif buffer
queue capacity.
* Continuously drain Rx memif queue while connection shutdown
is in progress. This avoids flooding the memif buffer while the
request to delete the connection is being processed.
* Rework FFmpeg MCM plugin to delete Rx connection just before
returning an error from read_packet(). This avoids memif errors due
to the delayed deleletion of the memif connection in read_close().

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

* Fix buffer context deletion flow

When the enqueue() function returns, the buffer context has to be
deleted even if the function returned an error. Change the flow to
avoid memory leaking.

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>

---------

Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>